Add Statement.getMultiple for retrieving multiple rows.
Using statement.get for retrieving rows is much slower than `each` and `all` due
to thread synchronization. By fetching multiple rows (but not all if there are huge
number of rows) we can decrease the synchronization cost per row.

Using getMultiple it is possible to implement Statement.each with support for
promises or node-streams with good performance.